购物车和结算完成

+++ spzx-modules/spzx-order/src/main/java/com/spzx/order/domain/OrderItem.java
@@ -46,6 +46,6 @@ public class OrderItem extends BaseEntity
     /** 商品购买的数量 */
     @Excel(name = "商品购买的数量")
     @Schema(description = "商品购买的数量")
-    private Long skuNum;
+    private Integer skuNum;
 
 }
